Switch Square type to u8
To save space, as opposed to usize, which was only being used to avoid
conversion for array access (hopefully the compiler does not do any
additional checks with "as usize" on a u8).  If anything, this
resulted in a small speedup around 1%.

+++ src/board/evaluate.rs
@@ -15,7 +15,7 @@ impl Board {
         let mut score = self.material[WHITE] - self.material[BLACK];
 
         for sq64 in self.bitboards[Piece::WP as usize].into_iter() {
-            score += PAWN_TABLE[sq64];
+            score += PAWN_TABLE[sq64 as usize];
 
             if self.bitboards[Piece::WP as usize].isolated_pawn(sq64) {
                 score += PAWN_ISOLATED_SCORE;
@@ -23,12 +23,12 @@ impl Board {
 
             if self.bitboards[Piece::BP as usize].passed_pawn(sq64, WHITE) {
                 // Todo: formalize RANKS for sq64
-                score += PAWN_PASSED_SCORE[sq64/8];
+                score += PAWN_PASSED_SCORE[(sq64/8) as usize];
             }
         }
 
         for sq64 in self.bitboards[Piece::BP as usize].into_iter() {
-            score -= PAWN_TABLE[MIRROR64[sq64]];
+            score -= PAWN_TABLE[MIRROR64[sq64 as usize]];
 
             if self.bitboards[Piece::BP as usize].isolated_pawn(sq64) {
                 score -= PAWN_ISOLATED_SCORE;
@@ -36,44 +36,44 @@ impl Board {
 
             if self.bitboards[Piece::WP as usize].passed_pawn(sq64, BLACK) {
                 // Todo: formalize RANKS for sq64
-                score -= PAWN_PASSED_SCORE[7 - sq64/8];
+                score -= PAWN_PASSED_SCORE[(7 - sq64/8) as usize];
             }
         }
 
         for sq64 in self.bitboards[Piece::WN as usize].into_iter() {
-            score += KNIGHT_TABLE[sq64];
+            score += KNIGHT_TABLE[sq64 as usize];
         }
 
         for sq64 in self.bitboards[Piece::BN as usize].into_iter() {
-            score -= KNIGHT_TABLE[MIRROR64[sq64]];
+            score -= KNIGHT_TABLE[MIRROR64[sq64 as usize]];
         }
 
         for sq64 in self.bitboards[Piece::WB as usize].into_iter() {
-            score += BISHOP_TABLE[sq64];
+            score += BISHOP_TABLE[sq64 as usize];
         }
 
         for sq64 in self.bitboards[Piece::BB as usize].into_iter() {
-            score -= BISHOP_TABLE[MIRROR64[sq64]];
+            score -= BISHOP_TABLE[MIRROR64[sq64 as usize]];
         }
 
         for sq64 in self.bitboards[Piece::WR as usize].into_iter() {
-            score += ROOK_TABLE[sq64];
+            score += ROOK_TABLE[sq64 as usize];
         }
 
         for sq64 in self.bitboards[Piece::BR as usize].into_iter() {
-            score -= ROOK_TABLE[MIRROR64[sq64]];
+            score -= ROOK_TABLE[MIRROR64[sq64 as usize]];
         }
 
         if self.material[BLACK] <= ENDGAME_MATERIAL {
-            score += KING_END_TABLE[self.king_sq[WHITE]];
+            score += KING_END_TABLE[self.king_sq[WHITE] as usize];
         } else {
-            score += KING_OPEN_TABLE[self.king_sq[WHITE]];
+            score += KING_OPEN_TABLE[self.king_sq[WHITE] as usize];
         }
 
         if self.material[WHITE] <= ENDGAME_MATERIAL {
-            score -= KING_END_TABLE[MIRROR64[self.king_sq[BLACK]]];
+            score -= KING_END_TABLE[MIRROR64[self.king_sq[BLACK] as usize]];
         } else {
-            score -= KING_OPEN_TABLE[MIRROR64[self.king_sq[BLACK]]];
+            score -= KING_OPEN_TABLE[MIRROR64[self.king_sq[BLACK] as usize]];
         }
 
         if self.side == WHITE {
